Mini Project/Code/requirements.txt
psycopg2-binary
sv-ttk
tkcalendar
ttkthemes
